+++ databento/common/data.py
@@ -48,7 +48,7 @@ MBO_MSG: List[Tuple[str, Union[type, str]]] = RECORD_HEADER + [
     ("order_id", np.uint64),
     ("price", np.int64),
     ("size", np.uint32),
-    ("flags", np.int8),
+    ("flags", np.uint8),
     ("channel_id", np.uint8),
     ("action", "S1"),  # 1 byte chararray
     ("side", "S1"),  # 1 byte chararray
@@ -62,7 +62,7 @@ MBP_MSG: List[Tuple[str, Union[type, str]]] = RECORD_HEADER + [
     ("size", np.uint32),
     ("action", "S1"),  # 1 byte chararray
     ("side", "S1"),  # 1 byte chararray
-    ("flags", np.int8),
+    ("flags", np.uint8),
     ("depth", np.uint8),
     ("ts_recv", np.uint64),
     ("ts_in_delta", np.int32),
@@ -93,7 +93,7 @@ DEFINITION_MSG: List[Tuple[str, Union[type, str]]] = RECORD_HEADER + [
     ("price_ratio", np.int64),
     ("inst_attrib_value", np.int32),
     ("underlying_id", np.uint32),
-    ("cleared_volume", np.int32),
+    ("_reserved1", "S4"),
     ("market_depth_implied", np.int32),
     ("market_depth", np.int32),
     ("market_segment_id", np.uint32),
@@ -102,11 +102,11 @@ DEFINITION_MSG: List[Tuple[str, Union[type, str]]] = RECORD_HEADER + [
     ("min_lot_size_block", np.int32),
     ("min_lot_size_round_lot", np.int32),
     ("min_trade_vol", np.uint32),
-    ("open_interest_qty", np.int32),
+    ("_reserved2", "S4"),
     ("contract_multiplier", np.int32),
     ("decay_quantity", np.int32),
     ("original_contract_size", np.int32),
-    ("reserved1", "S4"),
+    ("_reserved3", "S4"),
     ("trading_reference_date", np.uint16),
     ("appl_id", np.int16),
     ("maturity_year", np.uint16),
@@ -125,9 +125,9 @@ DEFINITION_MSG: List[Tuple[str, Union[type, str]]] = RECORD_HEADER + [
     ("underlying", "S21"),  # 21 byte chararray
     ("strike_price_currency", "S4"),
     ("instrument_class", "S1"),
-    ("reserved2", "S2"),
+    ("_reserved4", "S2"),
     ("strike_price", np.int64),
-    ("reserved3", "S6"),
+    ("_reserved5", "S6"),
     ("match_algorithm", "S1"),  # 1 byte chararray
     ("md_security_trading_status", np.uint8),
     ("main_fraction", np.uint8),
@@ -170,6 +170,20 @@ IMBALANCE_MSG: List[Tuple[str, Union[type, str]]] = RECORD_HEADER + [
     ("dummy", "S1"),
 ]
 
+STATISTICS_MSG: List[Tuple[str, Union[type, str]]] = RECORD_HEADER + [
+    ("ts_recv", np.uint64),
+    ("ts_ref", np.uint64),
+    ("price", np.int64),
+    ("quantity", np.int32),
+    ("sequence", np.uint32),
+    ("ts_in_delta", np.int32),
+    ("stat_type", np.uint16),
+    ("channel_id", np.uint16),
+    ("update_action", np.uint8),
+    ("stat_flags", np.uint8),
+    ("dummy", "S6"),
+]
+
 
 STRUCT_MAP: Dict[Schema, List[Tuple[str, Union[type, str]]]] = {
     Schema.MBO: MBO_MSG,
@@ -193,6 +207,7 @@ STRUCT_MAP: Dict[Schema, List[Tuple[str, Union[type, str]]]] = {
     Schema.OHLCV_1D: OHLCV_MSG,
     Schema.DEFINITION: DEFINITION_MSG,
     Schema.IMBALANCE: IMBALANCE_MSG,
+    Schema.STATISTICS: STATISTICS_MSG,
 }
 
 
@@ -208,20 +223,21 @@ DEFINITION_CHARARRAY_COLUMNS = [
     "security_type",
     "unit_of_measure",
     "underlying",
+    "strike_price_currency",
+    "instrument_class",
     "match_algorithm",
     "security_update_action",
     "user_defined_instrument",
-    "strike_price_currency",
 ]
 
 DEFINITION_PRICE_COLUMNS = [
     "min_price_increment",
-    "display_factor",
     "high_limit_price",
     "low_limit_price",
     "max_price_variation",
     "trading_reference_price",
     "min_price_increment_amount",
+    "price_ratio",
     "strike_price",
 ]
 
@@ -288,6 +304,13 @@ IMBALANCE_DROP_COLUMNS = [
     "dummy",
 ]
 
+STATISTICS_DROP_COLUMNS = [
+    "ts_recv",
+    "length",
+    "rtype",
+    "dummy",
+]
+
 DEFINITION_COLUMNS = [
     x
     for x in (np.dtype(DEFINITION_MSG).names or ())
@@ -296,6 +319,12 @@ DEFINITION_COLUMNS = [
 
 IMBALANCE_COLUMNS = [
     x for x in (np.dtype(IMBALANCE_MSG).names or ()) if x not in IMBALANCE_DROP_COLUMNS
+]
+
+STATISTICS_COLUMNS = [
+    x
+    for x in (np.dtype(STATISTICS_MSG).names or ())
+    if x not in STATISTICS_DROP_COLUMNS
 ]
 
 COLUMNS = {
@@ -333,4 +362,5 @@ COLUMNS = {
     Schema.OHLCV_1D: OHLCV_HEADER_COLUMNS,
     Schema.DEFINITION: DEFINITION_COLUMNS,
     Schema.IMBALANCE: IMBALANCE_COLUMNS,
+    Schema.STATISTICS: STATISTICS_COLUMNS,
 }