+++ databento/historical/api/timeseries.py
@@ -1,22 +1,25 @@
 from __future__ import annotations
 
-import warnings
 from datetime import date
-from io import BufferedIOBase, BytesIO
+from io import BufferedIOBase
+from io import BytesIO
 from os import PathLike
 from typing import List, Optional, Tuple, Union
 
 import pandas as pd
+
 from databento.common.dbnstore import DBNStore
 from databento.common.deprecated import deprecated
-from databento.common.enums import Compression, Dataset, Encoding, Schema, SType
-from databento.common.error import BentoWarning
-from databento.common.parsing import (
-    datetime_to_string,
-    optional_datetime_to_string,
-    optional_symbols_list_to_string,
-)
-from databento.common.validation import validate_enum, validate_semantic_string
+from databento.common.enums import Compression
+from databento.common.enums import Dataset
+from databento.common.enums import Encoding
+from databento.common.enums import Schema
+from databento.common.enums import SType
+from databento.common.parsing import datetime_to_string
+from databento.common.parsing import optional_datetime_to_string
+from databento.common.parsing import optional_symbols_list_to_string
+from databento.common.validation import validate_enum
+from databento.common.validation import validate_semantic_string
 from databento.historical.api import API_VERSION
 from databento.historical.http import BentoHttpAPI
 
@@ -29,35 +32,6 @@ class TimeSeriesHttpAPI(BentoHttpAPI):
     def __init__(self, key: str, gateway: str) -> None:
         super().__init__(key=key, gateway=gateway)
         self._base_url = gateway + f"/v{API_VERSION}/timeseries"
-
-    @deprecated
-    def stream(
-        self,
-        dataset: Union[Dataset, str],
-        start: Union[pd.Timestamp, date, str, int],
-        end: Optional[Union[pd.Timestamp, date, str, int]] = None,
-        symbols: Optional[Union[List[str], str]] = None,
-        schema: Union[Schema, str] = "trades",
-        stype_in: Union[SType, str] = "raw_symbol",
-        stype_out: Union[SType, str] = "instrument_id",
-        limit: Optional[int] = None,
-        path: Optional[Union[PathLike[str], str]] = None,
-    ) -> DBNStore:
-        """
-        The `.stream` method is deprecated and will be removed in a future version.
-        The method has been renamed to `.get_range`, which you can now use.
-        """
-        return self.get_range(
-            dataset=dataset,
-            start=start,
-            end=end,
-            symbols=symbols,
-            schema=schema,
-            stype_in=stype_in,
-            stype_out=stype_out,
-            limit=limit,
-            path=path,
-        )
 
     def get_range(
         self,
@@ -127,10 +101,12 @@ class TimeSeriesHttpAPI(BentoHttpAPI):
         stype_in_valid = validate_enum(stype_in, SType, "stype_in")
         symbols_list = optional_symbols_list_to_string(symbols, stype_in_valid)
         schema_valid = validate_enum(schema, Schema, "schema")
+        start_valid = datetime_to_string(start)
+        end_valid = optional_datetime_to_string(end)
         params: List[Tuple[str, Optional[str]]] = [
             ("dataset", validate_semantic_string(dataset, "dataset")),
-            ("start", datetime_to_string(start)),
-            ("end", optional_datetime_to_string(end)),
+            ("start", start_valid),
+            ("end", end_valid),
             ("symbols", symbols_list),
             ("schema", str(schema_valid)),
             ("stype_in", str(stype_in_valid)),
@@ -142,14 +118,6 @@ class TimeSeriesHttpAPI(BentoHttpAPI):
         # Optional Parameters
         if limit is not None:
             params.append(("limit", str(limit)))
-
-        self._pre_check_data_size(
-            symbols=symbols,
-            schema=schema_valid,
-            start=start,
-            end=end,
-            limit=limit,
-        )
 
         if path is not None:
             writer: BufferedIOBase = open(path, "x+b")
@@ -168,36 +136,6 @@ class TimeSeriesHttpAPI(BentoHttpAPI):
             return DBNStore.from_file(path)
         writer.seek(0)  # rewind for read
         return DBNStore.from_bytes(writer.read())
-
-    @deprecated
-    async def stream_async(
-        self,
-        dataset: Union[Dataset, str],
-        start: Union[pd.Timestamp, date, str, int],
-        end: Optional[Union[pd.Timestamp, date, str, int]] = None,
-        symbols: Optional[Union[List[str], str]] = None,
-        schema: Union[Schema, str] = "trades",
-        stype_in: Union[SType, str] = "raw_symbol",
-        stype_out: Union[SType, str] = "instrument_id",
-        limit: Optional[int] = None,
-        path: Optional[Union[PathLike[str], str]] = None,
-    ) -> DBNStore:
-        """
-        The `.stream_async` method is deprecated and will be removed in a future
-        version.
-        The method has been renamed to `.get_range_async`, which you can now use.
-        """
-        return await self.get_range_async(
-            dataset=dataset,
-            start=start,
-            end=end,
-            symbols=symbols,
-            schema=schema,
-            stype_in=stype_in,
-            stype_out=stype_out,
-            limit=limit,
-            path=path,
-        )
 
     async def get_range_async(
         self,
@@ -267,10 +205,12 @@ class TimeSeriesHttpAPI(BentoHttpAPI):
         stype_in_valid = validate_enum(stype_in, SType, "stype_in")
         symbols_list = optional_symbols_list_to_string(symbols, stype_in_valid)
         schema_valid = validate_enum(schema, Schema, "schema")
+        start_valid = datetime_to_string(start)
+        end_valid = optional_datetime_to_string(end)
         params: List[Tuple[str, Optional[str]]] = [
             ("dataset", validate_semantic_string(dataset, "dataset")),
-            ("start", datetime_to_string(start)),
-            ("end", optional_datetime_to_string(end)),
+            ("start", start_valid),
+            ("end", end_valid),
             ("symbols", symbols_list),
             ("schema", str(schema_valid)),
             ("stype_in", str(stype_in_valid)),
@@ -281,14 +221,6 @@ class TimeSeriesHttpAPI(BentoHttpAPI):
 
         if limit is not None:
             params.append(("limit", str(limit)))
-
-        self._pre_check_data_size(
-            symbols=symbols,
-            schema=schema_valid,
-            start=start,
-            end=end,
-            limit=limit,
-        )
 
         if path is not None:
             writer: BufferedIOBase = open(path, "x+b")
@@ -307,60 +239,3 @@ class TimeSeriesHttpAPI(BentoHttpAPI):
             return DBNStore.from_file(path)
         writer.seek(0)  # rewind for read
         return DBNStore.from_bytes(writer.read())
-
-    def _pre_check_data_size(  # noqa (prefer not to make static)
-        self,
-        symbols: Optional[Union[List[str], str]],
-        schema: Schema,
-        start: Optional[Union[pd.Timestamp, date, str, int]],
-        end: Optional[Union[pd.Timestamp, date, str, int]],
-        limit: Optional[int],
-    ) -> None:
-        if limit and limit < 10**7:
-            return
-
-        # Use heuristics to check ballpark data size
-        if (
-            _is_large_data_size_schema(schema)
-            or _is_greater_than_one_day(start, end)
-            or _is_large_number_of_symbols(symbols)
-        ):
-            warnings.warn(
-                message="The size of this streaming request is estimated "
-                "to be 5 GB or greater.\nWe recommend breaking your request "
-                "into smaller requests, or submitting a batch download request.\n"
-                "This warning can be suppressed: "
-                "https://docs.python.org/3/library/warnings.html",
-                category=BentoWarning,
-                stacklevel=3,  # This makes the error happen in user code
-            )
-
-
-def _is_large_number_of_symbols(symbols: Optional[Union[List[str], str]]) -> bool:
-    if not symbols:
-        return True  # Full universe
-
-    if isinstance(symbols, str):
-        symbols = symbols.split(",")
-
-    if len(symbols) >= 500:
-        return True
-
-    return False
-
-
-def _is_large_data_size_schema(schema: Schema) -> bool:
-    return schema in (Schema.MBO, Schema.MBP_10)
-
-
-def _is_greater_than_one_day(
-    start: Optional[Union[pd.Timestamp, date, str, int]],
-    end: Optional[Union[pd.Timestamp, date, str, int]],
-) -> bool:
-    if start is None or end is None:
-        return True
-
-    if pd.to_datetime(end) - pd.to_datetime(start) > pd.Timedelta(days=1):
-        return True
-
-    return False