+++ databento/live/dbn.py
@@ -1,6 +1,7 @@
 import asyncio
 import logging
 import queue
+import threading
 from concurrent import futures
 from typing import Callable, NewType, Optional, Union
 
@@ -21,6 +22,7 @@ DBNRecord = Union[
     databento_dbn.OHLCVMsg,
     databento_dbn.ImbalanceMsg,
     databento_dbn.InstrumentDefMsg,
+    databento_dbn.StatMsg,
     databento_dbn.SymbolMappingMsg,
     databento_dbn.SystemMsg,
     databento_dbn.ErrorMsg,
@@ -55,8 +57,9 @@ class DBNProtocol(asyncio.BufferedProtocol):
         self._buffer: bytearray = bytearray()
         self._client_callback = client_callback
         self._dbn_queue: Optional[DBNQueue] = None
-        self._decoder: databento_dbn.DbnDecoder = databento_dbn.DbnDecoder()
+        self._decoder: databento_dbn.DBNDecoder = databento_dbn.DBNDecoder()
         self._disconnected: "asyncio.Future[None]" = loop.create_future()
+        self._transport_lock = threading.Lock()
         self._transport = transport
 
     def __aiter__(self) -> "DBNProtocol":
@@ -90,7 +93,8 @@ class DBNProtocol(asyncio.BufferedProtocol):
                         "resuming reading with %d pending records",
                         self._dbn_queue.qsize(),
                     )
-                    self._transport.resume_reading()
+                    with self._transport_lock:
+                        self._transport.resume_reading()
 
         raise StopAsyncIteration()
 
@@ -118,7 +122,8 @@ class DBNProtocol(asyncio.BufferedProtocol):
                         "resuming reading with %d pending records",
                         self._dbn_queue.qsize(),
                     )
-                    self._transport.resume_reading()
+                    with self._transport_lock:
+                        self._transport.resume_reading()
 
         raise StopIteration()
 
@@ -171,21 +176,21 @@ class DBNProtocol(asyncio.BufferedProtocol):
             self._decoder.write(record_bytes)
         except ValueError:
             logger.critical("could not write to dbn decoder")
-            self._transport.close()
+            with self._transport_lock:
+                self._transport.close()
 
         try:
             records = self._decoder.decode()
         except ValueError:
             pass
         else:
-            for record, ts_out in records:
+            for record in records:
                 header = getattr(record, "hd", object())
                 ts_event = getattr(header, "ts_event", None)
                 logger.info(
-                    "decoded as %s record ts_event=%s ts_out=%s",
+                    "decoded as %s record ts_event=%s",
                     type(record).__name__,
                     ts_event,
-                    ts_out,
                 )
 
                 # Record Dispatch
@@ -207,4 +212,5 @@ class DBNProtocol(asyncio.BufferedProtocol):
                                 "record queue is full; %d record(s) to be processed",
                                 self._dbn_queue.qsize(),
                             )
-                            self._transport.pause_reading()
+                            with self._transport_lock:
+                                self._transport.pause_reading()